PID + Complementary Filter + It compiles!

+++ self_balancing_robot/self_balance/motors.cpp
@@ -32,9 +32,10 @@ void set_direction(int left, int right){
     debug_write("Set Direction");
 }
 
-void step(int count, float frequency){
+void step(int count, double frequency){
   if(frequency == 0) return;
-  float pulse_width = 1000.0 / (2 * frequency);
+  //frequency is a value between -20 and 20
+  double pulse_width = 1000.0 / (2 * frequency);
   Serial.println(pulse_width);//oh its in seconds. convert to milliseconds
     for(int i = 0; i < count; i++){
         digitalWrite(LEFT_MOTOR_STEP_PIN, HIGH);
